requirements.txt
Django==1.11.5
django-autofixture==0.12.1
djangorestframework==3.6.4
pytz==2017.2
